Create base rebase multi commit operation
