+++ app/src/ui/multi-commit-operation/reorder.tsx
@@ -1,8 +1,9 @@
-import { RebaseConflictState } from '../../lib/app-state'
 import { MultiCommitOperationKind } from '../../models/multi-commit-operation'
-import { BaseMultiCommitOperation } from './base-multi-commit-operation'
+import { BaseRebase } from './base-rebase'
 
-export abstract class Reorder extends BaseMultiCommitOperation {
+export abstract class Reorder extends BaseRebase {
+  protected conflictDialogOperationPrefix = 'reordering commits on'
+
   protected onBeginOperation = () => {
     const { repository, dispatcher, state } = this.props
     const { operationDetail } = state
@@ -21,65 +22,5 @@ export abstract class Reorder extends BaseMultiCommitOperation {
       lastRetainedCommitRef,
       true
     )
-  }
-
-  protected onContinueAfterConflicts = async (): Promise<void> => {
-    const {
-      repository,
-      dispatcher,
-      workingDirectory,
-      state,
-      conflictState,
-    } = this.props
-    const { operationDetail, targetBranch, originalBranchTip } = state
-
-    if (
-      conflictState === null ||
-      targetBranch === null ||
-      originalBranchTip === null ||
-      operationDetail.kind !== MultiCommitOperationKind.Reorder
-    ) {
-      this.endFlowInvalidState()
-      return
-    }
-
-    const { commits, currentTip } = operationDetail
-
-    await dispatcher.switchMultiCommitOperationToShowProgress(repository)
-
-    const rebaseConflictState: RebaseConflictState = {
-      kind: 'rebase',
-      currentTip,
-      targetBranch: targetBranch.name,
-      baseBranch: undefined,
-      originalBranchTip,
-      baseBranchTip: currentTip,
-      manualResolutions: conflictState.manualResolutions,
-    }
-
-    const rebaseResult = await dispatcher.continueRebase(
-      MultiCommitOperationKind.Reorder,
-      repository,
-      workingDirectory,
-      rebaseConflictState
-    )
-
-    return dispatcher.processMultiCommitOperationRebaseResult(
-      MultiCommitOperationKind.Reorder,
-      repository,
-      rebaseResult,
-      commits.length,
-      targetBranch.name
-    )
-  }
-
-  protected onAbort = async (): Promise<void> => {
-    const { repository, dispatcher } = this.props
-    this.onFlowEnded()
-    return dispatcher.abortRebase(repository)
-  }
-
-  protected onConflictsDialogDismissed = () => {
-    this.onInvokeConflictsDialogDismissed('reordering commits on')
   }
 }
