+++ app/src/ui/multi-commit-operation/squash.tsx
@@ -1,8 +1,9 @@
-import { RebaseConflictState } from '../../lib/app-state'
 import { MultiCommitOperationKind } from '../../models/multi-commit-operation'
-import { BaseMultiCommitOperation } from './base-multi-commit-operation'
+import { BaseRebase } from './base-rebase'
 
-export abstract class Squash extends BaseMultiCommitOperation {
+export abstract class Squash extends BaseRebase {
+  protected conflictDialogOperationPrefix = 'squashing commits on'
+
   protected onBeginOperation = () => {
     const { repository, dispatcher, state } = this.props
     const { operationDetail } = state
@@ -27,65 +28,5 @@ export abstract class Squash extends BaseMultiCommitOperation {
       commitContext,
       true
     )
-  }
-
-  protected onContinueAfterConflicts = async (): Promise<void> => {
-    const {
-      repository,
-      dispatcher,
-      workingDirectory,
-      state,
-      conflictState,
-    } = this.props
-    const { operationDetail, targetBranch, originalBranchTip } = state
-
-    if (
-      conflictState === null ||
-      targetBranch === null ||
-      originalBranchTip === null ||
-      operationDetail.kind !== MultiCommitOperationKind.Squash
-    ) {
-      this.endFlowInvalidState()
-      return
-    }
-
-    const { commits, currentTip } = operationDetail
-
-    await dispatcher.switchMultiCommitOperationToShowProgress(repository)
-
-    const rebaseConflictState: RebaseConflictState = {
-      kind: 'rebase',
-      currentTip,
-      targetBranch: targetBranch.name,
-      baseBranch: undefined,
-      originalBranchTip,
-      baseBranchTip: currentTip,
-      manualResolutions: conflictState.manualResolutions,
-    }
-
-    const rebaseResult = await dispatcher.continueRebase(
-      MultiCommitOperationKind.Squash,
-      repository,
-      workingDirectory,
-      rebaseConflictState
-    )
-
-    return dispatcher.processMultiCommitOperationRebaseResult(
-      MultiCommitOperationKind.Squash,
-      repository,
-      rebaseResult,
-      commits.length + 1,
-      targetBranch.name
-    )
-  }
-
-  protected onAbort = async (): Promise<void> => {
-    const { repository, dispatcher } = this.props
-    this.onFlowEnded()
-    return dispatcher.abortRebase(repository)
-  }
-
-  protected onConflictsDialogDismissed = () => {
-    this.onInvokeConflictsDialogDismissed('squashing commits on')
   }
 }
